requirements.txt
# Core
pytz==2019.3
Django==3.0.3
django-configurations==2.2
gunicorn==20.0.4

# For the persistence stores
psycopg2-binary==2.8.4
dj-database-url==0.5.0

# Rest apis
djangorestframework==3.11.0

# Developer Tools
ipdb==0.12.3
ipython==7.11.1
flake8==3.7.9

# Testing
pytest==5.4.1
pytest-django==3.9.0
pytest-mock==3.1.0
